unix_common dir added make resolver unix_common stuff resolver supports in osx version
